Add ZTree implementation
The ZTree is the data structure that will be used to model schemas and
configuration. Since state machines have the requirment that any child
state you enter you must maintain that you are in the parent state we
can easily use the zipper tree to walk up the lineage and see what all
the states we are in for a given sub state.

Future changes will start builiding the schema around the ZTree.