Introduce Configuration Module
This module will be the driving force behind calculating what the next
state configuration is going to be. It will also communicate back to the
machine any states that were entered/exited from doing the transitions
so that the machine can invoke their logic.

The first functions introduced here are `compound?` and `atomic?` which
will be used later in recursion to keep entering states until we have
entered atomic states on microstep.