Adds ZTree.find/2
The `find` API will receive a lot of use in the `Configuration` module
where we will need it to locate `State` nodes that are the targets of
`Transition`s.

The implementation of `find` could use some work to use the ZTree in a
more clever way to do the traversal, but since this is all internal
details this implementation will allow us to move forward.