Runs elixir tests as part of CI workflow
At a later point we should consider how we are going to run tests for
multiple different implementations. Also it would be a nice to have to
only run tests for the changed implementations. For now this sets us up
to have the elixir suite running as part of our PR process.